+++ mne/io/fiff/__init__.py
@@ -1,2 +1,2 @@
-from .raw import RawFIF
+from .raw import Raw
 from .raw import read_raw_fif